Rewrote Goal class
Given that we're putting the PDF itself in charge of the markdown
formatting, we don't need the formatted/unformatted distinction anymore.
But, some handling of dates as dates would be nice since Psych parses
them from YAML as `Date`s.

Also, related to YAML-parsing, use the non-folding form of block text,
so you can write "real" markdown in your yaml.